+++ YumikoBot/Modulos/Anilist.cs
@@ -38,9 +38,11 @@ namespace Discord_Bot.Modulos
                 {
                     usuario = msgUserInter.Result.Content;
                     if(msgUsuario != null)
-                        await msgUsuario.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            await msgUsuario.DeleteAsync("Auto borrado de Yumiko");
                     if(msgUserInter.Result != null)
-                    await msgUserInter.Result.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            await msgUserInter.Result.DeleteAsync("Auto borrado de Yumiko");
                 }
                 else
                 {
@@ -52,9 +54,11 @@ namespace Discord_Bot.Modulos
                     });
                     await Task.Delay(3000);
                     if (msgError != null)
-                        await msgError.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            await msgError.DeleteAsync("Auto borrado de Yumiko");
                     if (msgUsuario != null)
-                        await msgUsuario.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            await msgUsuario.DeleteAsync("Auto borrado de Yumiko");
                     return;
                 }
             }
@@ -205,7 +209,8 @@ namespace Discord_Bot.Modulos
                     {
                         var msg = await ctx.RespondAsync($"Error: {x.Message}").ConfigureAwait(false);
                         await Task.Delay(3000);
-                        await msg.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            await msg.DeleteAsync("Auto borrado de Yumiko");
                     }
                 }
             }
@@ -217,7 +222,8 @@ namespace Discord_Bot.Modulos
                     _ => await ctx.RespondAsync($"Error inesperado").ConfigureAwait(false),
                 };
                 await Task.Delay(3000);
-                await msg.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                    await msg.DeleteAsync("Auto borrado de Yumiko");
             }
         }
 
@@ -302,8 +308,11 @@ namespace Discord_Bot.Modulos
                         bool result = int.TryParse(msgElegir.Result.Content, out int elegido);
                         if (result && elegido > 0)
                         {
-                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            {
+                                await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                                await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            }
                             var datos = data.Data.Page.media[elegido - 1];
                             if ((datos.isAdult == "false") || (datos.isAdult == true && ctx.Channel.IsNSFW))
                             {
@@ -421,25 +430,32 @@ namespace Discord_Bot.Modulos
                                     Footer = funciones.GetFooter(ctx)
                                 });
                                 await Task.Delay(3000);
-                                await msg.DeleteAsync("Auto borrado de Yumiko");
+                                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                                    await msg.DeleteAsync("Auto borrado de Yumiko");
                             }
                         }
                         else
                         {
                             var msg = await ctx.RespondAsync($"Debes escribir un numero válido").ConfigureAwait(false);
                             await Task.Delay(3000);
-                            await msg.DeleteAsync("Auto borrado de Yumiko");
-                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            {
+                                await msg.DeleteAsync("Auto borrado de Yumiko");
+                                await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                                await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            }
                         }
                     }
                     else
                     {
                         var msg = await ctx.RespondAsync($"Tiempo agotado esperando el número").ConfigureAwait(false);
                         await Task.Delay(3000);
-                        await msg.DeleteAsync("Auto borrado de Yumiko");
-                        await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                        await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                        {
+                            await msg.DeleteAsync("Auto borrado de Yumiko");
+                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                        }
                     }
                 }
                 else
@@ -448,7 +464,8 @@ namespace Discord_Bot.Modulos
                     {
                         var msg = await ctx.RespondAsync($"Error: {x.Message}").ConfigureAwait(false);
                         await Task.Delay(3000);
-                        await msg.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            await msg.DeleteAsync("Auto borrado de Yumiko");
                     }
                 }
             }
@@ -460,7 +477,8 @@ namespace Discord_Bot.Modulos
                     _ => await ctx.RespondAsync($"Error inesperado").ConfigureAwait(false),
                 };
                 await Task.Delay(3000);
-                await msg.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                    await msg.DeleteAsync("Auto borrado de Yumiko");
             }
         }
 
@@ -536,8 +554,11 @@ namespace Discord_Bot.Modulos
                         bool result = int.TryParse(msgElegir.Result.Content, out int elegido);
                         if (result && elegido > 0)
                         {
-                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            {
+                                await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                                await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            }
                             var datos = data.Data.Page.media[elegido - 1];
                             if ((datos.isAdult == "false") || (datos.isAdult == true && ctx.Channel.IsNSFW))
                             {
@@ -629,26 +650,35 @@ namespace Discord_Bot.Modulos
                                     Color = new DiscordColor(0xFF0000),
                                     Footer = funciones.GetFooter(ctx)
                                 });
-                                await Task.Delay(3000);
-                                await msg.DeleteAsync("Auto borrado de Yumiko");
+                                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                                {
+                                    await Task.Delay(3000);
+                                    await msg.DeleteAsync("Auto borrado de Yumiko");
+                                }
                             }
                         }
                         else
                         {
                             var msg = await ctx.RespondAsync($"Debes escribir un numero válido").ConfigureAwait(false);
-                            await Task.Delay(3000);
-                            await msg.DeleteAsync("Auto borrado de Yumiko");
-                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            {
+                                await Task.Delay(3000);
+                                await msg.DeleteAsync("Auto borrado de Yumiko");
+                                await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                                await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            }
                         }
                     }
                     else
                     {
                         var msg = await ctx.RespondAsync($"Tiempo agotado esperando el número").ConfigureAwait(false);
-                        await Task.Delay(3000);
-                        await msg.DeleteAsync("Auto borrado de Yumiko");
-                        await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                        await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                        {
+                            await Task.Delay(3000);
+                            await msg.DeleteAsync("Auto borrado de Yumiko");
+                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                        }
                     }
                 }
                 else
@@ -656,8 +686,11 @@ namespace Discord_Bot.Modulos
                     foreach (var x in data.Errors)
                     {
                         var msg = await ctx.RespondAsync($"Error: {x.Message}").ConfigureAwait(false);
-                        await Task.Delay(3000);
-                        await msg.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                        {
+                            await Task.Delay(3000);
+                            await msg.DeleteAsync("Auto borrado de Yumiko");
+                        }
                     }
                 }
             }
@@ -668,8 +701,11 @@ namespace Discord_Bot.Modulos
                     "The HTTP request failed with status code NotFound" => await ctx.RespondAsync($"No se ha encontrado el anime `{anime}`").ConfigureAwait(false),
                     _ => await ctx.RespondAsync($"Error inesperado").ConfigureAwait(false),
                 };
-                await Task.Delay(3000);
-                await msg.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                {
+                    await Task.Delay(3000);
+                    await msg.DeleteAsync("Auto borrado de Yumiko");
+                }  
             }
         }
 
@@ -740,8 +776,11 @@ namespace Discord_Bot.Modulos
                         bool result = int.TryParse(msgElegir.Result.Content, out int elegido);
                         if (result && elegido > 0)
                         {
-                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            {
+                                await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                                await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            }
                             var datos = data.Data.Page.characters[elegido - 1];
                             string descripcion = datos.description;
                             descripcion = funciones.NormalizarDescription(funciones.LimpiarTexto(descripcion));
@@ -781,19 +820,25 @@ namespace Discord_Bot.Modulos
                         else
                         {
                             var msg = await ctx.RespondAsync($"Debes escribir un numero válido").ConfigureAwait(false);
-                            await Task.Delay(3000);
-                            await msg.DeleteAsync("Auto borrado de Yumiko");
-                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            {
+                                await Task.Delay(3000);
+                                await msg.DeleteAsync("Auto borrado de Yumiko");
+                                await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                                await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                            } 
                         }
                     }
                     else
                     {
                         var msg = await ctx.RespondAsync($"Tiempo agotado esperando el número").ConfigureAwait(false);
-                        await Task.Delay(3000);
-                        await msg.DeleteAsync("Auto borrado de Yumiko");
-                        await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
-                        await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                        {
+                            await Task.Delay(3000);
+                            await msg.DeleteAsync("Auto borrado de Yumiko");
+                            await elegirMsg.DeleteAsync("Auto borrado de Yumiko");
+                            await msgElegir.Result.DeleteAsync("Auto borrado de Yumiko");
+                        }
                     }
                 }
                 else
@@ -801,8 +846,11 @@ namespace Discord_Bot.Modulos
                     foreach (var x in data.Errors)
                     {
                         var msg = await ctx.RespondAsync($"Error: {x.Message}").ConfigureAwait(false);
-                        await Task.Delay(3000);
-                        await msg.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                        {
+                            await Task.Delay(3000);
+                            await msg.DeleteAsync("Auto borrado de Yumiko");
+                        }
                     }
                 }
             }
@@ -813,8 +861,11 @@ namespace Discord_Bot.Modulos
                     "The HTTP request failed with status code NotFound" => await ctx.RespondAsync($"No se ha encontrado el personaje `{personaje}`").ConfigureAwait(false),
                     _ => await ctx.RespondAsync($"Error inesperado, mensaje: [{ex.Message}").ConfigureAwait(false),
                 };
-                await Task.Delay(5000);
-                await msg.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                {
+                    await Task.Delay(5000);
+                    await msg.DeleteAsync("Auto borrado de Yumiko");
+                }
             }
         }
         
@@ -836,7 +887,8 @@ namespace Discord_Bot.Modulos
                         request.AddHeader("content-type", "application/json");
                         var procesando = await ctx.RespondAsync("Procesando imagen..").ConfigureAwait(false);
                         IRestResponse response = client.Execute(request);
-                        await procesando.DeleteAsync("Auto borrado de Yumiko");
+                        if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            await procesando.DeleteAsync("Auto borrado de Yumiko");
                         switch (response.StatusCode)
                         {
                             case HttpStatusCode.OK:
@@ -908,7 +960,8 @@ namespace Discord_Bot.Modulos
             {
                 DiscordMessage msgError = await ctx.RespondAsync(msg).ConfigureAwait(false);
                 await Task.Delay(3000);
-                await msgError.DeleteAsync("Auto borrado de Yumiko").ConfigureAwait(false);
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                    await msgError.DeleteAsync("Auto borrado de Yumiko").ConfigureAwait(false);
             }
         }
 
@@ -983,8 +1036,11 @@ namespace Discord_Bot.Modulos
                 {
                     _ => await ctx.RespondAsync($"Error inesperado: {ex.Message}").ConfigureAwait(false),
                 };
-                await Task.Delay(3000);
-                await msg.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                {
+                    await Task.Delay(3000);
+                    await msg.DeleteAsync("Auto borrado de Yumiko");
+                }
                 return;
             }
         }
