+++ YumikoBot/Modulos/NSFW.cs
@@ -55,8 +55,11 @@ namespace Discord_Bot.Modulos
             else
             {
                 var msg = await ctx.RespondAsync("No puedes hacerte eso a ti mismo");
-                await Task.Delay(3000);
-                await msg.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                {
+                    await Task.Delay(3000);
+                    await msg.DeleteAsync("Auto borrado de Yumiko");
+                }
             }
         }
 
@@ -81,8 +84,11 @@ namespace Discord_Bot.Modulos
             else
             {
                 var msg = await ctx.RespondAsync("No puedes hacerte eso a ti mismo");
-                await Task.Delay(3000);
-                await msg.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                {
+                    await Task.Delay(3000);
+                    await msg.DeleteAsync("Auto borrado de Yumiko");
+                }
             }
         }
 
@@ -360,16 +366,23 @@ namespace Discord_Bot.Modulos
             if (usuario1 == null || usuario2 == null)
             {
                 var msg = await ctx.RespondAsync("Debes mencionar a dos personas");
-                await Task.Delay(3000);
-                await msg.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                {
+                    await Task.Delay(3000);
+                    await msg.DeleteAsync("Auto borrado de Yumiko");
+                }
+                    
             }
             else
             {
                 if (ctx.User.Id == usuario1.Id || ctx.User.Id == usuario2.Id)
                 {
                     var msg = await ctx.RespondAsync("No puedes hacerte eso a ti mismo");
-                    await Task.Delay(3000);
-                    await msg.DeleteAsync("Auto borrado de Yumiko");
+                    if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                    {
+                        await Task.Delay(3000);
+                        await msg.DeleteAsync("Auto borrado de Yumiko");
+                    }
                 }
                 else
                 {
