+++ YumikoBot/Modulos/Usuarios.cs
@@ -78,8 +78,11 @@ namespace Discord_Bot.Modulos
             if (!msgFechaInter.TimedOut)
             {
                 bool result = funciones.ParsearFechaFromString(msgFechaInter.Result.Content, out DateTime fecha);
-                await msgFechaInter.Result.DeleteAsync("Auto borrado de Yumiko");
-                await msgFecha.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                {
+                    await msgFechaInter.Result.DeleteAsync("Auto borrado de Yumiko");
+                    await msgFecha.DeleteAsync("Auto borrado de Yumiko");
+                }
                 if (result)
                 {
                     var msgOcultar = await ctx.RespondAsync(embed: new DiscordEmbedBuilder
@@ -93,8 +96,11 @@ namespace Discord_Bot.Modulos
                         bool result2 = int.TryParse(msgOcultarInter.Result.Content, out int mostrarEdadInt);
                         if (result2)
                         {
-                            await msgOcultarInter.Result.DeleteAsync("Auto borrado de Yumiko");
-                            await msgOcultar.DeleteAsync("Auto borrado de Yumiko");
+                            if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                            {
+                                await msgOcultarInter.Result.DeleteAsync("Auto borrado de Yumiko");
+                                await msgOcultar.DeleteAsync("Auto borrado de Yumiko");
+                            }
                             switch (mostrarEdadInt)
                             {
                                 case 1:
@@ -129,8 +135,11 @@ namespace Discord_Bot.Modulos
             }
             if (msgError != null)
             {
-                await Task.Delay(3000);
-                await msgError.DeleteAsync("Auto borrado de Yumiko");
+                if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+                {
+                    await Task.Delay(3000);
+                    await msgError.DeleteAsync("Auto borrado de Yumiko");
+                } 
             }
         }
 
@@ -139,8 +148,11 @@ namespace Discord_Bot.Modulos
         {
             await usuariosService.DeleteBirthday(ctx);
             var msg = await ctx.RespondAsync("Cumpleaños borrado correctamente");
-            await Task.Delay(3000);
-            await msg.DeleteAsync("Auto borrado de Yumiko");
+            if (ctx.Channel.PermissionsFor(ctx.Guild.CurrentMember) == DSharpPlus.Permissions.ManageMessages)
+            {
+                await Task.Delay(3000);
+                await msg.DeleteAsync("Auto borrado de Yumiko");
+            }
         }
     }
 }
